requirements.txt
google-cloud-bigquery>=3.0.0
pandas>=1.3.0
db-dtypes
python-dotenv>=1.0.0
wfdb